requirements.txt
# Core dependencies for Pneumonia Detection Project
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
numpy>=1.24.0
pillow>=9.5.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
pandas>=2.0.0
tqdm>=4.65.0
opencv-python>=4.8.0

# For explainability
grad-cam>=1.4.8

# For visualization
seaborn>=0.12.0

# For model summary
torchinfo>=1.8.0

# For training progress
tqdm>=4.65.0

# Optional but recommended
jupyter>=1.0.0
ipywidgets>=8.0.0
